refactor: api 요청에 대한 request 비지니스 로직 customHook으로 이전
